moved helper.get_username to main
